+++ app/core/services/google/__init__.py
@@ -1,6 +1,6 @@
 """Google services package."""
 
-from app.core.services.google.sheets import GoogleSheetsService
-from app.core.services.google.drive import GoogleDriveService
+from app.core.services.google.sheets import SheetsService
+from app.core.services.google.drive import DriveService
 
-__all__ = ['GoogleSheetsService', 'GoogleDriveService'] 
+__all__ = ['SheetsService', 'DriveService']
