--- app/core/services/sheets.py
@@ -1,225 +1,0 @@
-"""Google Sheets service.
-
-Layer: core
-"""
-
-import os
-import logging
-from typing import List, Dict, Any, Optional, TypedDict, cast, Tuple
-from google.oauth2.credentials import Credentials
-from google.oauth2 import service_account
-from googleapiclient.discovery import build
-from googleapiclient.errors import HttpError
-
-logger = logging.getLogger(__name__)
-
-# Get TWO_WAY_SYNC_ENABLED from environment variable, default to False
-TWO_WAY_SYNC_ENABLED = os.environ.get("TWO_WAY_SYNC_ENABLED", "false").lower() == "true"
-
-def check_two_way_sync() -> None:
-    """Check if two-way sync is enabled.
-    
-    Raises:
-        RuntimeError: If two-way sync is not enabled.
-    """
-    if not TWO_WAY_SYNC_ENABLED:
-        raise RuntimeError("Two-way sync is not enabled. Set TWO_WAY_SYNC_ENABLED=true to enable.")
-
-class SheetResponse(TypedDict):
-    """Type for sheet response."""
-    spreadsheetId: str
-    updatedRange: str
-    updatedRows: int
-    updatedColumns: int
-    updatedCells: int
-
-class SheetMetadata(TypedDict):
-    """Type for sheet metadata."""
-    spreadsheetId: str
-    properties: Dict[str, Any]
-    sheets: List[Dict[str, Any]]
-    spreadsheetUrl: str
-
-class SheetsService:
-    """Google Sheets service for spreadsheet operations."""
-
-    def __init__(self, credentials: Optional[Credentials] = None) -> None:
-        """Initialize Sheets service.
-        
-        Args:
-            credentials: Optional Google credentials. If None, loads from GOOGLE_SVC_KEY.
-        """
-        if credentials is None:
-            key_path = os.environ.get("GOOGLE_SVC_KEY")
-            if not key_path:
-                raise ValueError("GOOGLE_SVC_KEY environment variable not set")
-            credentials = service_account.Credentials.from_service_account_file(
-                key_path,
-                scopes=['https://www.googleapis.com/auth/spreadsheets']
-            )
-        
-        self.service = build('sheets', 'v4', credentials=credentials)
-        self.spreadsheets = self.service.spreadsheets()
-        logger.info("Sheets service initialized")
-
-    def get_sheet_data(self, spreadsheet_id: str, range_name: str) -> list[list[str]]:
-        """Get data from a Google Sheet.
-        
-        Args:
-            spreadsheet_id: ID of the spreadsheet
-            range_name: A1 notation range to read
-            
-        Returns:
-            2D array of values
-        """
-        try:
-            result = self.service.spreadsheets().values().get(
-                spreadsheetId=spreadsheet_id,
-                range=range_name
-            ).execute()
-            return cast(list[list[str]], result.get("values", []))
-        except Exception as e:
-            logger.error(f"Error reading sheet: {e}")
-            raise
-
-    def update_sheet_data(
-        self, spreadsheet_id: str, range_name: str, values: list[list[str]]
-    ) -> dict[str, str]:
-        """Update data in a Google Sheet.
-
-        Args:
-            spreadsheet_id: ID of the spreadsheet
-            range_name: A1 notation range to update
-            values: 2D array of values to write
-            
-        Returns:
-            Update response
-        """
-        if not TWO_WAY_SYNC_ENABLED:
-            raise RuntimeError("two-way sync disabled")
-        try:
-            body = {
-                "values": values
-            }
-            result = self.service.spreadsheets().values().update(
-                spreadsheetId=spreadsheet_id,
-                range=range_name,
-                valueInputOption="RAW",
-                body=body
-            ).execute()
-            logger.info(f"Updated sheet {spreadsheet_id} range {range_name}")
-            return cast(dict[str, str], result)
-        except Exception as e:
-            logger.error(f"Error updating sheet: {e}")
-            raise
-
-    def append_sheet_data(
-        self, spreadsheet_id: str, range_name: str, values: list[list[str]]
-    ) -> dict[str, str]:
-        """Append data to a specific range in a spreadsheet.
-        
-        Args:
-            spreadsheet_id: ID of the spreadsheet
-            range_name: A1 notation range to append to
-            values: 2D array of values to append
-            
-        Returns:
-            Append response
-        """
-        if not TWO_WAY_SYNC_ENABLED:
-            raise RuntimeError("two-way sync disabled")
-        try:
-            body = {"values": values}
-            result = (
-                self.spreadsheets.values()
-                .append(
-                    spreadsheetId=spreadsheet_id,
-                    range=range_name,
-                    valueInputOption="RAW",
-                    insertDataOption="INSERT_ROWS",
-                    body=body,
-                )
-                .execute()
-            )
-            return cast(dict[str, str], result)
-        except HttpError as error:
-            raise Exception(f"Error appending sheet data: {error}")
-
-    def clear_sheet_data(self, spreadsheet_id: str, range_name: str) -> dict[str, str]:
-        """Clear data from a specific range in a spreadsheet.
-        
-        Args:
-            spreadsheet_id: ID of the spreadsheet
-            range_name: A1 notation range to clear
-            
-        Returns:
-            Clear response
-        """
-        if not TWO_WAY_SYNC_ENABLED:
-            raise RuntimeError("two-way sync disabled")
-        try:
-            result = (
-                self.spreadsheets.values()
-                .clear(spreadsheetId=spreadsheet_id, range=range_name)
-                .execute()
-            )
-            return cast(dict[str, str], result)
-        except HttpError as error:
-            raise Exception(f"Error clearing sheet data: {error}")
-
-    # ------------------------------------------------------------------
-    # Convenience helpers used by channel logic
-    # ------------------------------------------------------------------
-    def open_sheet(self, sheet_id: str) -> dict[str, str]:
-        """Return the spreadsheet metadata.
-        
-        Args:
-            sheet_id: ID of the spreadsheet
-            
-        Returns:
-            Spreadsheet metadata
-        """
-        return cast(dict[str, str], self.spreadsheets.get(spreadsheetId=sheet_id).execute())
-
-    def copy_template(
-        self, src_id: str, dst_title: str, folder_id: str
-    ) -> tuple[str, str]:
-        """Copy ``src_id`` to ``dst_title`` inside ``folder_id``.
-        
-        Args:
-            src_id: Source spreadsheet ID
-            dst_title: Destination spreadsheet title
-            folder_id: Destination folder ID
-            
-        Returns:
-            Tuple of (spreadsheet ID, URL)
-        """
-        if not TWO_WAY_SYNC_ENABLED:
-            raise RuntimeError("two-way sync disabled")
-        body = {"name": dst_title, "parents": [folder_id]}
-        copy = self.service.files().copy(fileId=src_id, body=body).execute()
-        sheet_id = copy["id"]
-        url = f"https://docs.google.com/spreadsheets/d/{sheet_id}"
-        return sheet_id, url
-
-    def append_rows(self, sheet_id: str, rows: list[list[str]]) -> dict[str, str]:
-        """Append ``rows`` to the first worksheet.
-        
-        Args:
-            sheet_id: ID of the spreadsheet
-            rows: 2D array of values to append
-            
-        Returns:
-            Append response
-        """
-        if not TWO_WAY_SYNC_ENABLED:
-            raise RuntimeError("two-way sync disabled")
-        body = {"values": rows}
-        result = self.spreadsheets.values().append(
-            spreadsheetId=sheet_id,
-            range="A1",
-            valueInputOption="USER_ENTERED",
-            insertDataOption="INSERT_ROWS",
-            body=body,
-        ).execute()
-        return cast(dict[str, str], result)
